pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perler-gen"
version = "0.1.0"
description = "Image to Perler bead pattern generator (MVP)."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Perler-Gen"}]
dependencies = [
  "Pillow>=10.0",
  "numpy>=1.24",
  "reportlab>=4.0",
  "scikit-image>=0.21",
  "scikit-learn>=1.3",
]

[project.optional-dependencies]
 dev = ["pytest>=7.4"]

[project.scripts]
perler-gen = "perler_gen.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
